+++ data-new/src/main/java/tv/anilibria/module/data/analytics/features/WebPlayerAnalytics.kt
@@ -13,7 +13,7 @@ class WebPlayerAnalytics(
     private val sender: AnalyticsSender
 ) {
 
-    fun open(from: String, releaseId: Int) {
+    fun open(from: String, releaseId: Long) {
         sender.send(
             AnalyticsConstants.web_player_open,
             from.toNavFromParam(),
